feat: separate mint card wrapper, pixel perfect
